% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSpatialObject.r
\name{addNewCoordinates}
\alias{addNewCoordinates}
\title{addNewCoordinates}
\usage{
addNewCoordinates(myData, myX, myY, myNewX = "X", myNewY = "Y",
  currentProj = NULL, newProj = NULL)
}
\arguments{
\item{myData}{A data.frame to be made spatial - must include columns of spatial data}

\item{myX}{Quoted name of column of easting or longitude coordinates}

\item{myY}{Quoted name of cColumn of northing or latitude coordinates}

\item{myNewX}{Quoted name of new column of easting or longitude coordinates}

\item{myNewY}{Quoted name of new column of northing or latitude coordinates}

\item{currentProj}{The current projection of the coordinate data as a character vector}

\item{newProj}{A new projection as a character vector.}
}
\description{
This function allows you to add new columns of coordinates in a different projection.
}
\examples{
addNewCoordinates(quakes,myX="long",myY="lat",myNewX = "easting",myNewY = "northing", currentProj = projWGS84, newProj = projNZTM)

}
\keyword{shapefile}
