% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.wheel.r
\name{col.wheel}
\alias{col.wheel}
\title{Colour Wheel}
\usage{
col.wheel(str, nearby = 3, cex = 0.75)
}
\arguments{
\item{str}{String of colour name}

\item{nearby}{Proximity of nearby colours. Defaults to 3.}

\item{cex}{Size of labels on piechart}
}
\description{
This function allows you to produce a colour wheel and list of matching colours based on the text provided and then add nearby colours
}
\examples{
col.wheel()
}
\keyword{colour}
