% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_to_shapefile.R
\name{convert.txt.shp}
\alias{convert.txt.shp}
\title{Convert text files to shapefiles}
\usage{
convert.txt.shp(text_file, filename, type)
}
\arguments{
\item{text_file}{Character string of the name of a text file. The text file must have the following columns: id, lat, lon}

\item{filename}{The desired output filename}

\item{type}{The desired output type of the shapefile. 5=Polygons, 3=Polylines, 1=Points}
}
\description{
This function allows you to convert text files to shapefiles
}
\examples{
convert.txt.shp()
}
\keyword{shapefiles}
\keyword{textfiles,}
