% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doug.rbind.list.r
\name{Doug.rbind.list}
\alias{Doug.rbind.list}
\title{doug.rbind.list}
\usage{
Doug.rbind.list(dfs)
}
\arguments{
\item{dfs}{List of dataframes}
}
\description{
Tfunction to rbind all the elements in a list of data frames
}
\examples{
Doug.rbind.list()
}
\keyword{rbind}
