% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DougSplit.R
\name{DougSplit}
\alias{DougSplit}
\title{DougSplit}
\usage{
DougSplit(mytext, myPart = 1, split = "\\\\.", ...)
}
\arguments{
\item{mytext}{character vector, each element of which is to be split. Other inputs, including a factor, will give an error.}

\item{myPart}{Identify the part of the split to return. Default = 1.}

\item{split}{character vector (or object which can be coerced to such) containing regular expression(s) (unless fixed = TRUE) to use for splitting. If empty matches occur, in particular if split has length 0, x is split into single characters. If split has length greater than 1, it is re-cycled along x.}
}
\description{
Wrapper for base::strsplit
}
\examples{
DougSplit()
}
\keyword{rbind}
