% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_bars.r
\name{errorbarsBarplot}
\alias{errorbarsBarplot}
\title{Error bars for barplots}
\usage{
errorbarsBarplot(yv, z, nn)
}
\arguments{
\item{yv}{height of the bars}

\item{z}{length of the error bars (+/-)}

\item{nn}{labels for the x-axis}
}
\description{
Error bar function for bar plots from The R Book (Page 56)
}
\examples{
errorbarsBarplot()
}
\keyword{barplot}
\keyword{errorbar}
