% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_bars.r
\name{errorbarsScatterplot}
\alias{errorbarsScatterplot}
\title{Y Error bars for scatterplots}
\usage{
errorbarsScatterplot(x, y, upper, lower = upper, length = 0.1, ...)
}
\arguments{
\item{x}{vector of x-values}

\item{y}{vector of y-values}

\item{upper}{vector of upper limits for errorbars}

\item{lower}{vector of lower limits for errorbars. Defaults to upper}

\item{length}{width of cross bars on error bars}
}
\description{
error bar function from http://monkeysuncle.stanford.edu/?p=485
}
\examples{
errorbarsScatterplot()

}
\keyword{errorbar}
\keyword{scatterplot}
