% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportPlot.R
\name{ExportPlot}
\alias{ExportPlot}
\title{Export ggplot}
\usage{
ExportPlot(gplot, filename, width = 2, height = 1.5)
}
\arguments{
\item{gplot}{ggplot2 object}

\item{filename}{character string}

\item{width}{width of plot in inches}

\item{height}{height of plot in inches}
}
\description{
This function exports a ggplot2 object to a pdf, eps and png file. Code credited to https://gist.github.com/sheymann/2399659
}
\examples{
is.outlier()
}
\keyword{eps}
\keyword{ggplot2}
\keyword{pdf}
\keyword{png}
