% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDropboxPath.r
\name{getDropboxPath}
\alias{getDropboxPath}
\title{Identify Dropbox path}
\usage{
getDropboxPath()
}
\description{
This function identifies the path to a Dropbox folder
}
\examples{
getDropboxPath()
}
\keyword{dropbox}
