% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random forest outputs.r
\name{GetImportance}
\alias{GetImportance}
\title{Get importance}
\usage{
GetImportance(myModel)
}
\arguments{
\item{myModel}{Model object from a random forest model}
}
\description{
This function allows you to calculate the importance variables
}
\examples{
GetImportance()

}
\keyword{forest}
\keyword{random}
