% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random forest outputs.r
\name{GetObsPreds}
\alias{GetObsPreds}
\title{Get Observed and Predicted data}
\usage{
GetObsPreds(myModel)
}
\arguments{
\item{myModel}{Model object from a random forest model}
}
\description{
This function allows you to produce a frame of observed and predicted values
}
\examples{
GetObsPreds()
}
\keyword{forest}
\keyword{random}
