% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random forest outputs.r
\name{GetRegression}
\alias{GetRegression}
\title{Get Regression outputs}
\usage{
GetRegression(myDataFrame, myY, myX, myMethod = lm)
}
\arguments{
\item{myDataFrame}{Dataframe of data for regression model.}

\item{myY}{Name of response variable.}

\item{myX}{Vector of predictor variable names.}

\item{myMethod}{The type of regression model to run. Default = "lm".}
}
\description{
This function allows you to run a regression model
}
\examples{
GetRegression()
}
\keyword{forest}
\keyword{random}
