% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShapefile.R
\name{getShapefile}
\alias{getShapefile}
\title{getShapefile}
\usage{
getShapefile(myShapefile, myDir)
}
\arguments{
\item{myShapefile}{A character string with the shapefile name. Must not end in .shp}

\item{myDir}{The directory path where the shapefile sits. Must end with trailing slash}
}
\description{
This function allows you to open a shapefile using a filename and path
}
\examples{
getShapefile()
}
\keyword{shapefile}
