% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.outlier.r
\name{is.outlier}
\alias{is.outlier}
\title{Identify outliers}
\usage{
is.outlier(x)
}
\arguments{
\item{x}{Numeric vector}
}
\description{
This function produces a logical vector identifing outliers in a vector based on Davies, P.L. and Gather, U. (1993)."The identification of multiple outliers" (with discussion) J. Amer. Statist. Assoc., 88, 782-801.
}
\examples{
is.outlier()
}
\keyword{outlier}
