% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latticePanels.r
\name{LatticeSettings}
\alias{LatticeSettings}
\title{Doug's default lattice settings}
\usage{
LatticeSettings(myCols = 1:8, mypch = 15, mylty = 1, mylwd = 1,
  ChangeAxis = T)
}
\arguments{
\item{myCols}{vector or single value of colour data. Can be r colours or hex codes. Default = 1:8}

\item{mypch}{single value or vector of pch values. Default = 15}

\item{mylty}{set line type. Default = 1}

\item{mylwd}{set line width. Default = 1}

\item{ChangeAxis}{Default = T}
}
\description{
Doug's default lattice settings
}
\examples{


}
\keyword{correlation}
\keyword{lattice}
