% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSpatialObject.r
\name{makeSpatialObject}
\alias{makeSpatialObject}
\title{makeSpatialObject}
\usage{
makeSpatialObject(myData, myX, myY, currentProj = NULL, newProj = NULL)
}
\arguments{
\item{myData}{A data.frame to be made spatial - must include columns of spatial data}

\item{myX}{Column of easting or longitude coordinates}

\item{myY}{Column of northing or latitude coordinates}

\item{currentProj}{The current projection of the coordinate data as a character vector}

\item{newProj}{A new projection as a character vector. Use if you wish to reproject to a different coordinate system.}
}
\description{
This function allows you to create a SpatialPointsDataFrame from a data.frame and project (and potentially reproject) it to a specified projection.
}
\examples{
makeSpatialObject(quakes,myX="long",myY="lat",currentProj = projWGS84)

}
\keyword{shapefile}
