% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{packages}
\alias{packages}
\title{Packages}
\usage{
packages(x)
}
\arguments{
\item{x}{The desired package}
}
\description{
This function tries to load a library. If the library does not exist, it will install it for you.
}
\examples{
packages()
}
\keyword{install.packages}
\keyword{library,}
\keyword{require,}
