% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latticePanels.r
\name{panel.cor}
\alias{panel.cor}
\title{Correlation panel for lattice plots}
\usage{
panel.cor(x, y, digits = 2, prefix = "", cex.cor, ...)
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector}
}
\description{
put (absolute) correlations on pairs plot, with size proportional to the correlations.
}
\examples{

pairs(iris,lower.panel=panel.smooth, upper.panel=panel.cor)

}
\keyword{correlation}
\keyword{lattice}
