% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latticePanels.r
\name{panel.NZmapMG}
\alias{panel.NZmapMG}
\title{plot map of NZ coastline in NZMG}
\usage{
panel.NZmapMG(x, y, ...)
}
\arguments{
\item{x}{x variables}

\item{y}{y variables}
}
\description{
plot map of NZ coastline in NZMG
}
\examples{


}
\keyword{NZMG}
\keyword{lattice}
\keyword{map}
\keyword{panel}
