% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latticePanels.r
\name{panel.NZmapTM}
\alias{panel.NZmapTM}
\title{plot map of NZ coastline in NZTM}
\usage{
panel.NZmapTM(x, y, ...)
}
\arguments{
\item{x}{x variables}

\item{y}{y variables}
}
\description{
plot map of NZ coastline in NZTM
}
\examples{


}
\keyword{NZTM}
\keyword{lattice}
\keyword{map}
\keyword{panel}
