% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrettyCut.R
\name{PrettyCut}
\alias{PrettyCut}
\title{Pretty cut}
\usage{
PrettyCut(myFac)
}
\arguments{
\item{myFac}{Factor}
}
\description{
This function allows you to produce a tidy cut of breaks for nice legend plotting
}
\examples{
PrettyCut()
}
\keyword{legend}
