% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{projNZMG}
\alias{projNZMG}
\title{New Zealand Map Grid projection}
\format{A character string containing the CRS projection}
\source{
\url{}
}
\usage{
projNZMG
}
\description{
New Zealand Map Grid projection
}
\keyword{datasets}
