% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.columns.r
\name{rbind.columns}
\alias{rbind.columns}
\title{Rbind dataframes together}
\usage{
\method{rbind}{columns}(x, y, type = "all")
}
\arguments{
\item{x}{Dataframe 1}

\item{y}{Dataframe 2}

\item{type}{The type of rbind. all: rbinds the dataframes using all columns from both frames, filling any blank spaces with NA. match: rbinds the dataframes together, keeping only those columns that match.}
}
\description{
This function allows you to rbind dataframes together even if the columns don't match.
}
\examples{
rbind.columns()
}
\keyword{rbind}
