% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproject.R
\name{reproject}
\alias{reproject}
\title{Reproject a shapefile}
\usage{
reproject(myShapefile, myProjection = projNZTM)
}
\arguments{
\item{myShapefile}{The filenameshapefile of interest}

\item{myProjection}{The desired projection as a character string. Default is projNZTM}
}
\description{
This function changes the spatial projection of a shapefile
}
\examples{
reproject()
}
