% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp2raster.R
\name{shp2raster}
\alias{shp2raster}
\title{Convert shapefiles to rasters}
\usage{
shp2raster(shp, mask.raster, label, value, bkg = 0, transform = FALSE,
  proj.from = NA, proj.to = NA, map = TRUE, save = TRUE, mask = TRUE)
}
\arguments{
\item{shp}{The shapefile you wish to convert}

\item{mask.raster}{A raster with the desired resolutions and extent of the output raster}

\item{label}{A string giving the desired filename for the raster (if map=TRUE) and the raster name}

\item{value}{An integer value for cells in the raster that are within the shapefile}

\item{bkg}{An integer value for cells in the raster outside the shapefile. Defaults to 0.}

\item{transform}{Do you wish to transform the projection of the shapefile prior to conversion to a raster. Defaults to FALSE.}

\item{proj.from}{The projection of the shapefile. Only used if transform = TRUE.}

\item{proj.to}{The desired projection of the raster. Only used if transform = TRUE.}

\item{map}{Do you want to output a map to the plot window at the end of conversion? Defaults to TRUE}

\item{save}{Do you wish to save a copy of the raster to the working directory? Defaults to TRUE. Filename is based on label.}

\item{mask}{Do you want the output raster to be masked by the mask.raster? Defaults to TRUE}
}
\description{
This function allows you to convert a shapefile to a raster.
}
\examples{
shp2raster()
}
\keyword{raster}
\keyword{shapefile,}
