% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_bars.r
\name{x.error.bars}
\alias{x.error.bars}
\title{X Error bars for scatterplots}
\usage{
x.error.bars(x, y, xbar)
}
\arguments{
\item{x}{vector of x-values}

\item{y}{vector of y-values}

\item{xbar}{vector of errors for x values}
}
\description{
X Error bars for scatterplots
}
\examples{
x.error.bars()

}
\keyword{errorbar}
\keyword{scatterplot}
