% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_bars.r
\name{xy.error.bars}
\alias{xy.error.bars}
\title{X & Y Error bars for scatterplots}
\usage{
xy.error.bars(x, y, xbar, ybar)
}
\arguments{
\item{x}{vector of x-values}

\item{y}{vector of y-values}

\item{xbar}{vector of errors for x values}

\item{ybar}{vector of errors for y values}
}
\description{
error bar function from http://monkeysuncle.stanford.edu/?p=485
}
\examples{
errorbarsScatterplot()

}
\keyword{errorbar}
\keyword{scatterplot}
