% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_bars.r
\name{y.error.bars}
\alias{y.error.bars}
\title{Y Error bars for scatterplots}
\usage{
y.error.bars(x, y, ybar, xlabel, ylabel)
}
\arguments{
\item{x}{vector of x-values}

\item{y}{vector of y-values}

\item{ybar}{vector of errors for y values}

\item{xlabel}{label for x-axis}

\item{ylabel}{label for y-axis}
}
\description{
Y Error bars for scatterplots
}
\examples{
y.error.bars()

}
\keyword{errorbar}
\keyword{scatterplot}
